requirements.txt
apache-beam[gpc]==2.55.0
psycopg2-binary==2.9.9
google-cloud-storage
pyarrow
setuptools
apache_beam[dataframe]
pandas
fsspec
gcsfs